make sed rate gen from matrix

at height height[i], the sedimentation rate is specified by the pdf approxfun(sedrate, matrix[i,])

Usage

sed_rate_from_matrix(height, sedrate, matrix, rate = 1, expand_domain = TRUE)

Arguments

height

vector of heights

sedrate

vector of sed. rates x values

matrix

matrix of sed rate y values

rate

numeric, rate of the Poisson point process determining frequency of sedimentation rate changes.

expand_domain

should sedimentation rates be defined below/above the highest/lowest height in the section? If TRUE, the sed rate values are the values at the closest interpolated point, if FALSE it will be NA

Value

a function factory for usage with sedrate_to_multiadm

See also